backend/requirements.txt
# AviLingo Backend Dependencies
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==41.0.7
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Redis for caching and rate limiting
redis==5.0.1

# Validation and settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client (for external APIs)
httpx==0.26.0

# Cloud Storage (S3/MinIO)
boto3==1.34.25
botocore==1.34.25

# Background Tasks
celery==5.3.6
kombu==5.3.4

# Audio Processing (for Whisper)
openai-whisper==20231117
ffmpeg-python==0.2.0

# Date/time utilities
python-dateutil==2.8.2

# Development
python-dotenv==1.0.0

